restructure files (initial commit)
