+++ src/common/linked_list.h
