+++ src/common/queue.c
@@ -1,4 +1,4 @@
-#include "inc/queue.h"
+#include "queue.h"
 
 #include <assert.h>
 #include <stdlib.h>
