+++ src/common/type.h
