+++ src/entity/player.h
@@ -1,8 +1,8 @@
 #ifndef _PLAYER_H_
 #define _PLAYER_H_
 
+#include "../common/type.h"
 #include "entity.h"
-#include "window.h"
 
 #include <stdbool.h>
