+++ src/game/level.c
@@ -1,4 +1,4 @@
-#include "inc/level.h"
+#include "level.h"
 
 /**
  * TODO:
