+++ src/game/time.c
@@ -1,4 +1,4 @@
-#include "inc/time.h"
+#include "time.h"
 
 #include <sys/time.h>
 #include <stddef.h>
