+++ src/gfx/stars.h
