+++ src/gfx/window.h
@@ -1,0 +1,10 @@
+#ifndef _WINDOW_H_
+#define _WINDOW_H_
+
+#define WINDOW_WIDTH		640
+#define WINDOW_HEIGHT		640
+#define WINDOW_BUFFER       250
+#define WINDOW_FLAGS		0
+#define RENDERER_FLAGS		SDL_RENDERER_ACCELERATED
+
+#endif
